Fix use after free in qiskit iterator layer
This commit fixes a use after free issue in the rust layer where the
arrays allocated by the qiskit c api would be freed before we actually
returned them to the user from the iterator. Since everything is done as
shared references we lost the data and potentially got garbage out when
accessing the iterator fields. This fixes it so we only free on dropping
or the subsequent iteration step so the data is always preserved for
it's lifetime.