Add basic error infra and return opaque Job handle from sampler run.
This includes a rename of some functions to be prefixed with qkrt_ (the ones
we will likely keep around in the actual library API eventually), and a new
qkrt_job_free function used to free the Job instance returned during job
submit.

The Job struct is exposed from C as an opaque construct. The idea is we
would provide a pointer to this to any function that needs to query job
status etc.