Support multiple instances.
Searching for a backend now searches all instances of the account,
unless the 'instance' field is set in the user config file, in
which case only that instance is used.

Currently, backend search ignores errors when attempting to list
backends, since it may be the case that the user does not have
permission to list for a particular instance. We should improve
this so that if a specific 'instance' is set, it's a hard error
and otherwise only ignore if it's an access issue.