Add least busy and instance info APIs.
Function qkrt_backend_search_results_least_busy can be used to return
the backend from that query with the fewest jobs in the queue. This
is defined at the search result level rather than at the backend
level (as is done in the Python client) since it better matches the
C data ownership model we have, and is also more flexible since when
we add filtering capabilities users can get the least busy backend
from the list of filtered results.

It's also now possible to use qkrt_backend_instance_crn and
qkrt_backend_instance_name to get the CRN and human-readable
names of a Backend. The test_job_submit.c demo program has been
updated to show the source instance name next to each listed backend.