Remove debug file write (#2)
In the initial bring up of the job submission workflow to figure out if
the generated job payload was valid or not the client code was
configured to write out the request payload body to a file in /tmp. This
was primarily done for debugging to verify the correct payload was
generated and check the payload. However after the client was working
the validation here was not needed anymore, but removing it was
overlooked. This has been fully superseded by the logging mechanism and
doesn't provide general debugging. This commit fixes this oversight and
removes the file write.