Add `const` markers to C declarations and `QkCircuit`
This adds the `const` markers that I thought should be there.  I
modified a bit of Rust code, but perhaps I don't understand some way in
which the `QkCircuit` might be modified.  On the other hand, if I am to
trust the Rust compiler, then I must not have missed anything. :)