Log warnings on unknown gate types instead of panicking
This commit fixes the processing when encountering a custom gate that is
outside Qiskit's standard operations when generating a target for an IBM
backend. Several IBM backends can contain details for custom operations
specific to a given backend. In Python these are defined with custom
operation classes to define the semantics of the operation and then
added to the target. The Qiskit C API doesn't yet offer this level of
flexibility and can't represent these custom operations in the target.
During the bring up of this library for testing encountering these
custom operations would just panic which was fine for testing, but is
not desirable behavior when actually using the library. This commit
switches to logging a warning whenever a backend's properties indicate
it supports gate or operation that Qiskit's C API can't represent
currently. It will no longer crash the program when trying to build the
target in these cases.